requirements.txt
amqp==2.5.2
asgiref==3.2.7
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
bcrypt==3.1.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.7.0rc1
Django==3.0.6
django-bootstrap-datepicker-plus==3.0.5
django-celery-beat==2.0.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-querysetsequence==0.12
django-storages==1.9.1
django-timezone-field==4.0
flake8==3.8.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
kombu==4.6.8
mccabe==0.6.1
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
redis==3.5.2
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
vine==1.3.0
zope.interface==5.1.0
zxcvbn==4.4.28
